feat: 컨텍스트 관심사 분리
